push Taxes files to git
